refactor: settingsディレクトリを作成しconfig.pyを移動
